refactor: unify hooks directory
